idx type Long으로 변경 production #41

+++ src/main/java/com/tsp/new_tsp_front/api/production/domain/FrontProductionEntity.java
@@ -29,7 +29,7 @@ public class FrontProductionEntity extends NewCommonMappedClass {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "idx")
-    private Integer idx;
+    private Long idx;
 
     @Column(name = "title")
     @NotEmpty(message = "프로덕션 제목 입력은 필수입니다.")
